ai_backend/requirements.txt
# FastAPI Backend Requirements
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# AI and ML dependencies
huggingface-hub>=0.20.0
transformers>=4.30.0
torch>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.20

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# HTTP and networking
requests>=2.31.0
aiohttp>=3.8.0

# Async support
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0

# Logging and monitoring
python-json-logger>=2.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # For testing FastAPI endpoints